// sluzbenik/client/src/app/model/obavestenje.ts
export interface Obavestenje {
    id?: string;
    nazivOrgana: string;
    sedisteOrgana: string;
    brojPredmeta: string;
    datumPodnosenja: string;
    ime: string;
    prezime: string;
    adresa: string;
    opisInformacije: string;
    datumUvida: string;
    vremeUvida: string;
    od: string;
    do: string;
    mestoUvida: string;
    ulicaUvida: string;
    brojUlice: string;
    brojKancelarije: string;
    trosak: string
}