// sluzbenik/client/src/app/model/zahtev.ts
export interface Zahtev {
    id: string;
    nazivOrgana: string;
    sediste: string;
    svrha: string;
    nacinDostave: string;
    opisInformacije: string;
    mestoPredaje: string;
    datum: string;
    ime: string;
    prezime: string;
    adresa: string;
    kontakt: string;
    status: string;
}